\name{s525-package}
\alias{s525-package}
\alias{s525}
\docType{package}
\title{
\packageTitle{s525}
}
\description{
\packageDescription{s525}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{s525}
\packageIndices{s525}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{s525}

Maintainer: \packageMaintainer{s525}
}
\references{
% Literature or other references for background information
}
% Optionally other standard keywords, one per line, from file KEYWORDS in
% the R documentation directory
\keyword{ package }
\examples{
# simple examples of the most important functions
}
